Nearly done last episode, just needs exercie.

+++ code/cysum.pyx
@@ -1,7 +1,7 @@
-cpdef float sumnum(int n):
-  cdef float result=0.0
+cpdef double sumnum(int n):
+  cdef double result=0.0
   cdef int i
   for i in range(n):
-    result+=float(i)
+    result+=i
   return result
